Refactor paths on screens.js
