--- wchat/screens/Home/styles.js
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { mainBackgroundColor } from '../../shared/utils/colors.js'
-
-const styles = StyleSheet.create({
-  tabbar: {
-   backgroundColor: mainBackgroundColor,
- },
-})
-
-export default styles
