+++ wchat/screens/shared/components/SearchButton/styles.js
